Fix linker warning
